fixed big when saving minMaxSearch radius

+++ viewer/viewercorebase.cpp
@@ -74,7 +74,7 @@ void ViewerCoreBase::setCommonViewerOptions()
 
 	m_OptionsMap->setPropertyAs<bool>("propagateZooming", false);
 	m_OptionsMap->setPropertyAs<bool>("showLables", false );
-	m_OptionsMap->setPropertyAs<uint8_t>("minMaxSearchRadius", 20);
+	m_OptionsMap->setPropertyAs<uint16_t>("minMaxSearchRadius", 20);
 	//logging
 	m_OptionsMap->setPropertyAs<uint16_t>("logDelayTime", 5000);
 	m_OptionsMap->setPropertyAs<bool>("showErrorMessages", true );
